echogpt/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
scikit-learn
joblib
numpy
PyYAML